requirements.txt
Django==2.1.1
django-cors-headers==3.4.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.1.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
text-unidecode==1.3
